Bink

+++ HCNV2/HCNV2.cpp
@@ -4,14 +4,14 @@
 #include<iomanip>
 using namespace std;
 
-bool Giao(int, int, int, int, int, int, int, int, int&, int&, int&, int&);
-string DienTichGiao(int, int, int, int);
+bool Giao(long long, long long, long long, long long, long long, long long, long long, long long, long long&, long long&, long long&, long long&);
+string DienTichGiao(long long, long long, long long, long long);
 
 int main()
 {
-	int xA, yA, xB, yB, xC, yC, xD, yD;
+	long long xA, yA, xB, yB, xC, yC, xD, yD;
 	cin >> xA >> yA >> xB >> yB >> xC >> yC >> xD >> yD;
-	int L, R, T, B;
+	long long L, R, T, B;
 	if (Giao(xA, yA, xB, yB, xC, yC, xD, yD, L, R, T, B))
 		cout << DienTichGiao(L, R, T, B);
 	else
@@ -19,10 +19,10 @@ int main()
 	return 0;
 }
 
-bool Giao(int xA, int yA, int xB, int yB, int xC, int yC, int xD, int yD, int& L, int& R, int& T, int& B)
+bool Giao(long long xA, long long yA, long long xB, long long yB, long long xC, long long yC, long long xD, long long yD, long long& L, long long& R, long long& T, long long& B)
 {
-	int L1 = min(xA, xB), R1 = max(xA, xB), T1 = max(yA, yB), B1 = min(yA, yB);
-	int L2 = min(xC, xD), R2 = max(xC, xD), T2 = max(yC, yD), B2 = min(yC, yD);
+	long long L1 = min(xA, xB), R1 = max(xA, xB), T1 = max(yA, yB), B1 = min(yA, yB);
+	long long L2 = min(xC, xD), R2 = max(xC, xD), T2 = max(yC, yD), B2 = min(yC, yD);
 	if (R1 <= L2 || R2 <= L1 || T1 <= B2 || T2 <= B1)
 		return false;
 	else
@@ -35,10 +35,10 @@ bool Giao(int xA, int yA, int xB, int yB, int xC, int yC, int xD, int yD, int& L
 	return true;
 }
 
-string DienTichGiao(int L, int R, int T, int B)
+string DienTichGiao(long long L, long long R, long long T, long long B)
 {
 	stringstream stream;
-	int c1 = abs(L - R), c2 = abs(T - B);
+	long long c1 = abs(L - R), c2 = abs(T - B);
 	stream << c1 * c2;
 	return stream.str();
 }
